Upgrade jQuery from 1.8.3 to 2.1.1
NOTE: This is a major upgrade. It should work just fine for all modern
browsers. The most significant thing that this will break is support for
older IE browsers. If you still need to support these browsers you can
override the `jquery_src` block and supply an version from the 1.x line.